fix logic

+++ Controllers/ProductController.cs
@@ -233,6 +233,13 @@ namespace vpp_server.Controllers
                     return NotFound(new ResponseDto { IsSuccess = false, Message = "Product not found" });
                 }
 
+                // Check if the product is part of any order
+                var isProductInOrder = await _context.OrderDetails.AnyAsync(od => od.ProductId == id);
+                if (isProductInOrder)
+                {
+                    return BadRequest(new ResponseDto { IsSuccess = false, Message = "Product cannot be deleted as it is part of an order" });
+                }
+
                 _context.Products.Remove(product);
                 await _context.SaveChangesAsync();
